requirements.txt
pytest==7.4.3
allure-pytest==2.13.5
allure-python-commons==2.13.5
requests==2.32.2
